+++ lib/services/followed-teams/model/followed_club.dart
